Feat/contest implementation

+++ server/src/main/java/com/asap/server/domain/CodeBattleMatch.java
@@ -54,7 +54,7 @@ public class CodeBattleMatch {
   private LocalDateTime created_at;
 
   @Column(name = "ai_order")
-  private int aiOrder;
+  private Long aiOrder;
 
   public CodeBattleMatch(CodeBattleContest contest, Users user1, Users user2, Users winner, String log) {
     this(contest, user1, user2, winner, null, log);
